% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartociudad_get_route.R
\name{cartociudad_get_route}
\alias{cartociudad_get_route}
\title{Driving and walking directions from Cartociudad API}
\usage{
cartociudad_get_route(latlon.orig, latlon.dest, vehicle = "car")
}
\arguments{
\item{latlon.orig}{Latitude and longitude of the starting point}

\item{latlon.dest}{Latitude and longitude of the destination point}

\item{vehicle}{Either \code{car} or \code{walking}}
}
\value{
A list containing the fields described in Cartociudad API
  documentation (see the link below).
}
\description{
Cartociudad API provides driving and walking routes between two
  points. This function quieries the API and provides the user the data in
  convenient form.
}
\examples{
res <- cartociudad_get_route(c(39.48,-0.37),
   c(39.484336,-0.358171),
   vehicle = "car")

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Carlos J. Gil Bellosta
}
