% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cartociudad_area.R
\name{get_cartociudad_area}
\alias{get_cartociudad_area}
\title{Area calculation}
\usage{
get_cartociudad_area(latitude, longitude, radius)
}
\arguments{
\item{latitude}{Point latitude in geographical coordinates (e.g., 40.3930144)}

\item{longitude}{Point longitude in geographical coordinates (e.g.,
-3.6596683)}

\item{radius}{Distance in meters (e.g., 500)}
}
\value{
A dataframe with the polygon that describes the area.
}
\description{
Returns the polygon that describes the area
}
\details{
This function calculates the area given a point and a radius in
  meters
}
\examples{
get_cartociudad_area(40.3930144, -3.6596683, 500)

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Luz Frias
}
