% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartociudad_geocode.R
\name{cartociudad_geocode}
\alias{cartociudad_geocode}
\title{Interface to Cartociudad geolocation API}
\usage{
cartociudad_geocode(full_address, on.error = "fail", ...)
}
\arguments{
\item{full_address}{Character string providing the full address to be
geolocated; e.g., "calle miguel servet 5, zaragoza". Adding the country may
cause problems.}

\item{on.error}{Defaults to \code{fail}; in such case, in case of errors in the API call, the process will fail. Set it to
"warn" and, in case of errors, the function will return \code{NULL} and a warning.}

\item{...}{Other parameters for the API. See Details section below.}
}
\value{
A data frame consisting of a single row per guess. See the reference
  below for an explanation of the data frame columns.
}
\description{
Geolocation of Spanish addresses via Cartociudad API calls, providing the
  full address in a single text string via \code{full_address}. It is
  advisable to add the street type (calle, etc.) and to omit the country
  name.
}
\details{
The entity geolocation API admits more parameters beyond the address field such as \code{id} or \code{type}. 
  You can use these extra arguments (see the References or the Examples sections below for further information) 
  at your own risk.
}
\examples{
# standard usage
res <- cartociudad_geocode(full_address = "plaza de cascorro 11, 28005 madrid")

#' # km 41 of A-23 motorway
res <- cartociudad_geocode("A-23 41")

# specific usage (see References for details)
res <- cartociudad_geocode("A-23 41", type = "portal", id = "600000000045", portal = 41)

# vectorized call
\dontrun{
addresses <- paste("A-23", 1:10)
res <- lapply(addresses, cartociudad_geocode, on.error = "warn")
}
}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Carlos J. Gil Bellosta
}
