% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartociudad_geocode.R
\name{cartociudad_geocode}
\alias{cartociudad_geocode}
\title{Interface to Cartociudad geolocation API}
\usage{
cartociudad_geocode(full_address, output_format = "JSON")
}
\arguments{
\item{full_address}{Character string providing the full address to be
geolocated; e.g., "calle miguel servet 5, zaragoza". Adding the country may
cause problems.}

\item{output_format}{Character string. Output format of the query: "JSON" or
"GeoJSON".}
}
\value{
A data frame consisting of a single row per query. See the reference
  below for an explanation of the data frame columns.
}
\description{
Geolocation of Spanish addresses via Cartociudad API calls, providing the
  full address in a single text string via \code{full_address}. It is
  advisable to add the street type (calle, etc.) and to omit the country
  name.
}
\examples{
# Query a single address
address <- "plaza de cascorro 11, 28005 madrid"
my.address <- cartociudad_geocode(full_address = address)
print(my.address)

# Query multiple addresses
address <- c(address, "plaza del ayunamiento 1, valencia")
my.address <- cartociudad_geocode(full_address = address)
print(my.address)

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Carlos J. Gil Bellosta
}
