% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartociudad_reverse_geocode.R
\name{cartociudad_reverse_geocode}
\alias{cartociudad_reverse_geocode}
\title{Reverse geocoding of locations}
\usage{
cartociudad_reverse_geocode(latitude, longitude)
}
\arguments{
\item{latitude}{Point latitude in geographical coordinates (e.g., 40.473219)}

\item{longitude}{Point longitude in geographical coordinates (e.g.,
-3.7227241)}
}
\value{
A data frame consisting of a single row per query. See the reference
  below for an explanation of the data frame columns.
}
\description{
Returns the address details of a geographical point in Spain.
}
\details{
This function performs reverse geocoding of a location. It returns
  the details of the closest address in Spain.
}
\examples{
# Query one point
cartociudad_reverse_geocode(40.473219, -3.7227241)

# Query multiple points
cartociudad_reverse_geocode(c(40.473219, 39.46979), c(-3.7227241, -0.376963))

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Luz Frias
}
