% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartociudad_reverse_geocode.R
\name{cartociudad_reverse_geocode}
\alias{cartociudad_reverse_geocode}
\title{Reverse geocoding of locations}
\usage{
cartociudad_reverse_geocode(latitude, longitude, ntries = 10)
}
\arguments{
\item{latitude}{Point latitude in geographical coordinates (e.g., 40.473219)}

\item{longitude}{Point longitude in geographical coordinates (e.g.,
-3.7227241)}

\item{ntries}{Numeric. In case of connection failure, number of \code{GET}
requests to be made before stopping the function call.}
}
\value{
A data frame consisting of a single row per query, with columns:
\item{tipo}{type of location.}
\item{tipo.via}{road type.}
\item{nombre.via}{road name.}
\item{num.via}{road number.}
\item{num.via.id}{internal id of this address in cartociudad database.}
\item{municipio}{town.}
\item{provincia}{province.}
\item{cod.postal}{zip code.}
}
\description{
Returns the address details of a geographical point in Spain.
}
\details{
This function performs reverse geocoding of a location. It returns
  the details of the closest address in Spain.
}
\examples{
# Query one point
cartociudad_reverse_geocode(40.473219, -3.7227241)

# Query multiple points
cartociudad_reverse_geocode(c(40.473219, 39.46979), c(-3.7227241, -0.376963))

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Luz Frias
}
