% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cartociudadmap.R
\name{get_cartociudadmap}
\alias{get_cartociudadmap}
\title{Get a Cartociudad Map}
\usage{
get_cartociudadmap(center, radius, add.censal.section = FALSE,
   add.postcode.area = FALSE, height = 600, width = 600)
}
\arguments{
\item{center}{a pair of numbers (latitude and longitude of the center of the
map)}

\item{radius}{approximate map "width" in kilometers}

\item{add.censal.section}{whether to add the limit of censal sections and
districts to the base map; note that this layer may not be available at low
zoom levels}

\item{add.postcode.area}{whether to add the limit of postal code areas to
the base map; note that this layer may not be available at low
zoom levels}

\item{height}{map height in pixels}

\item{width}{map width in pixels}
}
\value{
An object of class \code{ggmap} and \code{raster} which can be used
  within the \code{ggmap}framework.
}
\description{
Downloads static maps using Cartociudad API. These maps can be
  then plotted by functions such as \code{ggmap}.
}
\details{
This function, similar to \code{get_googlemap} or
  \code{get_openstreetmap} downloads a map from Cartociudad API and creates a
  \code{ggmap} compatible version of it.
}
\examples{
\dontrun{
  soria <- cartociudad_geocode("ayuntamiento soria")
  soria_map <- get_cartociudadmap(c(soria$lat, soria$lng), 1)
  ggmap::ggmap(soria_map)
}

}
\references{
\url{http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf}
}
\author{
Carlos J. Gil Bellosta
}
